Grocer list
Changes made include formatting with interpolation instead of concat. 
Removed empty parens on methods as unnecessary in ruby.
Need to use the hash rocket to use the push method.
Need to create the list total amount based on user input and a method that replicates pushing how many items the user wants.